[Pipeline] Docker version update broke manifest creation process
`docker pull` does not automatically pull anymore the image named, but
requests the current platform. If you don't supply the extra flag
`--platform`, then you get this:

```
$ docker pull yottadb/${REPO}:${DOCKERHUBTAG}-aarch64
Error response from daemon: no matching manifest for linux/amd64 in the
manifest list entries: no match for platform in manifest: not found
```

This commit adds an explicit `--platform`.

The previous version of Docker was v20, the new version is v29.

See this MR for why we needed to upgrade our Docker version:
https://gitlab.com/YottaDB/Util/YDBDevOps/-/merge_requests/238

This commit also reworks `docker login` per the recommandations in
https://forum.gitlab.com/t/docker-login-works-for-stage-build-but-fails-for-stage-deploy-in-the-same-pipeline/84915/2.

Later, while testing, `--provenance=false` turned out also to be needed to
the docker build step due to errors with manifests
`docker.io/my-repo/my-image-arm64:<some-tag> is a manifest list`. This
is also due to new docker versions, and the manifest commands being
considered experimental. See a full discussion here: https://stackoverflow.com/questions/75521775/buildx-docker-image-claims-to-be-a-manifest-list.

An earlier commit of this work was tested, and here's an example of
a successful run: https://gitlab.com/shabiel/YDB/-/pipelines/2161604565.