[#1180] Fixed dollar_ecode persistence after ydb_ci[p]_t bug.
background
----------
`ydb_ci_t` and `ydb_cip_t` are used to call M routines from C code with a multi threaded environment.
As they do essentially the exact same thing with different input format, the pair will be referred to as `ydb_ci[p]_t`.
In the code, the information about Ecode is stored in the `dollar_ecode` and `dollar_stack`.

Issue
------------
Previously, when a call to `ydb_ci[p]_t` would set Ecode, that would
put mumps permanently in an error state, even after the function returns.
This is not the correct behavior, and so this MR makes sure that the `dollar_ecode` variable and `dollar_stack`
are emptied when the function returns.

coding details
--------------
After `ydb_ci[p]_t` calls `threaded_api_ydb_engine_lock`,
the values of `dollar_ecode` and `dollar_stack` are nullified with the NULLIFY_DOLLAR_ECODE macro so that you cannot start in an error state.
At the end of the function call, the values of `dollar_ecode` and `dollar_stack` are again emptied to ensure that an error does not persist after the function is done.

Issue link
------------
the original issue is at https://gitlab.com/YottaDB/DB/YDB/-/issues/1180