[#933] Fix running JOB command with dash parameters no longer produces an error.
Background
----------------
In mumps the J[OB] command can be used to run another label concurrently.
There is also syntax for setting the command line of this JOB, `JOB nextLabel:cmd="value"`.
This can be useful if you want to pass information to the JOBed process through the value of `$zcmdline`.
The command line of a JOBed process is not intended to affect the behavior in any way, it just sets the value of `$zcmdline`.

Previous bug
-----------------
Before this MR, when you tried to run a JOB command where the command line has
parameters with a dash, you would incorrectly get an error. Now, you will no longer get an error.
Example mumps code below
```
 JOB nextLabel:cmd="-meaninglessparam"
 q
nextLabel
 W "hi",!
 W $zcomandline,!
```
Running this code would have produced an error in the JOB command causing it to exit.
You would still complete the parent process, only the new JOB would exit with an error.
So, running the above program would not result in "hi" or `$zcomandline` getting printed.

New behavior
------------------
The new behavior is simply that now the JOBed process will run regardless of command line.
So for the program above, "hi" will be printed and `$zcomandline` will print
"-meaninglessparam".

Coding details
-------------------
The file `sr_unix/cli_parse.c` is where the command line is parsed.
The alias `IS_JOBBED_PROCESS`, defined in `sr_unix/jobsp.h` checks for a command line argument or its negated version.
It used to be that when the function encountered a command line argument that it did not recognize it would add a message to the error buffer.
Now, it only does that if the program was not invoked by a JOB command.
Similarly, if `parse_arg()` finds an unrecognized arg, it will only report a failure if the program was not invoked by a JOB command.
Several files have been changed to use `IS_JOBBED_PROCESS` as well, as they had the code written out.

sr_unix/gtm_main.c
--------------------------
This function was producing an error if it found that it was invoked in direct mode and certain conditions are met,
unless it turned out that it was invoked through a job command.
It now uses `IS_JOBBED_PROCESS`.

sr_unix/jobchild_init.c
----------------------------
The `jobchild_init` function in this file works differently if it is in a JOB invocation.
It now uses `IS_JOBBED_PROCESS`.

sr_unix/ojchildparms.c
------------------------------
This file was checking if it was in a JOBed process, and now uses `IS_JOBBED_PROCESS`.
Additionally, it was at the same time assigning the value of the global variable sp through
hard to read lazy && logic, the readability of this logic has been improved.

Possible future work
---------------------------
There is still a bug where no error is given when you run direct mode without
starting with a dashed argument, for example, either of these.
```
$ydb_dist/yottadb -dir dummy -port
$ydb_dist/yottadb -dir dummy
```
However, you do get an error if you run direct mode starting with a dashed argument. For example this.
```
$ydb_dist/yottadb -dir -port
```
These cases should give the same result. Most likely, neither one should give an error.