[#858] YottaDB syslog messages will now print the calling function name (or libyottadb.so base address) + offset
Background
-----------------
Send_msg_va is a function that would, among other things, print the address of the function that called it.
Although, as it would be called by one of 2 wrapper functions in the same file,
it would actually look 2 calls back.
It would print the symbolic address where it was called, looking like this.
```
-- generated from 0x00007FC2DD9695CC.
```
This is not as useful as it could be.
It is hard to figure out where in the code a symbolic address is pointing to,
especially if you are reading that address off someone else's computer.
The usefulness of this message has been improved in this commit.

Change
----------
 It will now try to print the symbolic name of the function, for example
```
-- generated from 0x00007FC2DD9695CC (mupip_main+0x0000031a).
```
In the above message,
mupip_main means that the function `mupip_main()` is 2 calls up the stack from send_msg_va
0x31a means that the call happened hex 31a=decimal 794 bytes from the start of the function.

However, in most cases the symbolic name will be unavailable as we only export
a limited set of symbols from libyottadb.so that users can safely invoke.
In these cases, the backtrace will be given in entry point + offset.
```
generated from 00007F3B7FAB43E7 (0x7F600000+0x004B43E7)
```
0x7F600000 is the address of the shared object, in this case libyottadb.so
0x004B43E7 is the offset between libyottadb.so and the place where the function is called.
This way, you can use the offset to figure out where 00007F3B7FAB43E7 is pointing.

Coding details send_msg_va
--------------------------
The `dladdr` function can be used to get the values of the backtrace.
First, you use `backtrace()` to get the absolute values of the addresses of the backtraces.
On some systems(example:OpenSUSE), `backtrace()` will include itself and on other systems(example:Ubuntu) it will not.
To get around this, use the `dladdr` function to check if the first function in the backtrace
has libyottabd.so in the pathname of the shared object the function is in.
Then, call `dladdr` on the third function up from this base function.
`dladdr` will  give you the function name if it exists,
otherwise setting that field to NULL.
If `dladdr` sets the function name to NULL,
then the code will instead look at the address of the shared library
containing the address as that is also passed into `dladdr`.
By subtracting the backtrace address from the address of its shared library, you get the offset.
This information also could have been gotten with backtrace_symbols(), but that would have needlessly
given information for all the functions in the backtrace,
so `dladdr` is better here,
more information about not using backtrace_symbols() https://gitlab.com/YottaDB/DB/YDB/-/merge_requests/1799#note_2966789879

Coding details caller_id
------------------------
Additionally, this commit changes the `caller_id` function.
This function is used to get the absolute address of a recent point in the backtrace.
As in the `send_msg_va` function, the backtrace now adjusts for the different
behavior on different systems that do or do not include the backtrace function itself.
Additionally, this function was written in a way where if you tried to get the backtrace 3 or more steps back
you would access past the end of the backtrace array.
`caller_id` has been modified to `malloc` enough space to retrieve the backtrace.
Note: there was a strange behavior where when you call the backtrace function in `caller_id`
to get 6 steps of the backtrace(7 if backtrace function included), ASan reports an error.

Changes in sr_port/region_freeze.c
---------------------------------------------
In the YDBTest for this MR, there is a test that uses addresses from DBFREEZEON and DBFREEZEOFF syslog messages
to track down the position in the code that the DBFREEZEON and DBFREEZEOFF happened.
As this test cares about the exact wording of the line that creates the syslog message,
an informational comment has been added to both these places in the code
noting that changing these lines requires changing them in the test also.