[#917] Fix mupip dumpfhead, integ, upgrade, set, rundown with an empty -file argument
The first bug
--------------
if you invoke mupip dumpfhead, integ, upgrade, set, or rundown with an empty -file argument,
you would also get the same incorrect message saying a region or a file is valid input when only a file would be accepted.
```
>$gmt_dist/mupip set -FILE
File or Region:
```
For mupip dumpfhead and integ, you get the same message even if you don't include the -file argument at all
```
>$gmt_dist/mupip dumpfhead
File or Region:
```

The fix
-------
Now, the messages in `sr_unix/mupip_cmd.c` has been changed from `File or Region:` to `File:`
to reflect this for all 5 mentioned function, mupip dumpfhead, integ, upgrade, set, and rundown

The second bug
---------------
When you run mupip dumpfhead, a newline was not being printed until after the result.
```
>$gmt_dist/mupip dumpfhead
File or Region:fake.txt%DUMPFHEAD-F-NOTAPATH Path DEFAULT does not exist

%YDB-E-MUNOFINISH, MUPIP unable to finish all requested actions
```
Additionally, you could get a bad address when writing to the terminal error.
```
>$gtm_dist/mupip dumpfhead > x.out
%YDB-E-TERMWRITE, Error writing to terminal
%SYSTEM-E-ENO14, Bad address
```

Second bug cause
--------------
The second bug was happening because the newline was not getting printed immediately.
Instead, it would be added to the output buffer.
Normally, new stuff to print would be added to the buffer also, maintaining order.
However, mupip dumpfhead works with a call to the `system` function looking something like this.
```
"%s/mumps -run %%XCMD 'do dumpfhead^%%DUMPFHEAD(\"%s\")'"
```
The `system` function ignores the buffer, so the results were being printed in the wrong order.
Additionally, if the output device was not a terminal, the code would still try to print to the terminal.
This was causing the second part of this error.

Second bug fix
--------------
To fix this bug, call `iott_flush(io_std_device.out)` to flush the buffer before the system call.
Additionally, to fix the `TERMWRITE` error, add a check to the calls to  `iott_flush` to make them not happen if it's not flushing a terminal device.
The code now also checks that it is writing to a terminal before `iott_flush` in `iott_readfl`.
The check in `iott_readfl` is to ensure that `iott_flush` is not called if the output device is not a terminal when writing the prompt before the `iott_flush` call.
The code now also checks that it is writing to a terminal inside `iott_flush`.
As `iott_flush` is only intended to be called on a terminal device, as assert has been added to assert that this check is unnecessary.